It compiles

+++ xxd_model.sh
@@ -4,8 +4,13 @@
 #put model file in the working directory of this script
 #this script must be ran from the root directory of the project (so that the model data file is at main/model_data.cpp)
 
+if [ ! -n "$1" ]; then
+    echo "Please specify the model file name (e.g. model.vtflite)"
+    exit 1
+fi
+
 rm main/model_data.cpp
-echo -e '#include "model_data.h"\n'
+echo '#include "data.h"\n' > main/model_data.cpp
 xxd -i $1 >> main/model_data.cpp
-vname = $(echo "$1" | tr '.' '_')
-echo -e "\nconst uint8_t* model_data = ${vname};"
+vname=$(echo "$1" | tr '.' '_')
+echo "\nconst uint8_t* model_data = ${vname};" >> main/model_data.cpp
